feat: ✨ Redirect to absolute URLs for better caching (#363)
* feat: :sparkles: Add basic recirects to app and api

Use same mechanics as libnpmdiff to resolve specs and then see if it's any different from what we started with, if it is, we redirect to that URL instead. Intention of this feature is to improve caching by always pointing to absolute urls in cases where we can

* fix: :pencil2: Fix missed import in lib/npm-diff

* fix: :boom: Stop pre-building the examples.

Pre-building the examples was a nice thing, but I like the idea of having examples using "non-canonical" urls

* feat: :zap: Re-adding prebuilding of examples.

* test: Add testing for absolute specs

Since there's so many different scenarios that we want to redirect, I can't really think any better way than adding tests

* ci(actions): 🚀 Run tests on ci

* chore(vscode): Add commitscope actions

* refactor: Rename and clarify `parts-to-specs` -> `split-parts`

* docs: Document what `convertVersionsToSpecs` does

* refactor: ✨ Create new well-tested "immutable-spec" method.

Seemed a lot simpler to test all possible permutations if the method was designed to resolve single specs rather than doing it in pairs. Then we can test the pairs as well, but we don't have to worry about every single permutation.

* refactor: Refactor code for resolving immutable specs

* refactor: Refactor weird npmdiff method to destination-specs.

Instead of having weird method that EITHER does nothing and redirects, OR does everything, we now have a reusable method to get canonical immutable specs with redirect info, and then if there's no redirect, we only do diff at that point.

* fix: ✏ Fix oopsie in imports

* refactor: Stop using static pages, make main page handle, query for options

* style: Clean up React typings

* refactor: Refactor api endpoint for new structure

* refactor: Cleanup some imports in ...parts.tsx

* fix: Fix redirect for main page to avoid double params

* feat: Make sure everything is cached