feat: Set default glob in UI.
So to make URLs nicer and to make building a link to the tool excternally is useful. Noticed this when linking to the tool from an external source and realising the diff was useful. To opt out of the default glob, you can still specify the query param without value.