Fix tooltip going from uncontrolled to controlled state which is not allowed
We want the tooltip to disappear once we've clicked the button. We used to use the "controlled" state to set it to `false` once clicked, but radix prints out warnings when going from a "uncontrolled" (meaning it uses internal state management) to "controlled" (meaning we control the open state) component.

So instead we just conditionally completely re-render the button without the tooltip once clicked.